feat(django): Adapt South to Django 1.8
- Adapt South to support Django 1.6 through 1.8
- Merged migrate/syncdb behavior into unified command which understands how to apply schema
- Remove support for data fixtures from syncdb (loaddata/dumpdata)
