+++ src/south/db/generic.py
@@ -7,7 +7,10 @@ from django.core.management.color import no_style
 from django.db import transaction, models
 from django.db.utils import DatabaseError
 from django.db.backends.util import truncate_name
-from django.db.backends.creation import BaseDatabaseCreation
+try:
+    from django.db.backends.creation import BaseDatabaseCreation
+except ImportError:
+    from django.db.backends.base.creation import BaseDatabaseCreation
 from django.db.models.fields import NOT_PROVIDED
 from django.dispatch import dispatcher
 from django.conf import settings
@@ -28,6 +31,7 @@ except ImportError:
             res = instance.__dict__[self.func.__name__] = self.func(instance)
             return res
 
+from south.constants import DJANGO_17
 from south.logger import get_logger
 from south.utils.py3 import string_types, text_type
 
@@ -1012,9 +1016,10 @@ class DatabaseOperations(object):
         """
         if self.dry_run:
             self.pending_transactions += 1
-        transaction.commit_unless_managed(using=self.db_alias)
-        transaction.enter_transaction_management(using=self.db_alias)
-        transaction.managed(True, using=self.db_alias)
+        if not DJANGO_17:
+            transaction.commit_unless_managed(using=self.db_alias)
+            transaction.enter_transaction_management(using=self.db_alias)
+            transaction.managed(True, using=self.db_alias)
 
     def commit_transaction(self):
         """
@@ -1024,7 +1029,8 @@ class DatabaseOperations(object):
         if self.dry_run:
             return
         transaction.commit(using=self.db_alias)
-        transaction.leave_transaction_management(using=self.db_alias)
+        if not DJANGO_17:
+            transaction.leave_transaction_management(using=self.db_alias)
 
     def rollback_transaction(self):
         """
@@ -1034,7 +1040,8 @@ class DatabaseOperations(object):
         if self.dry_run:
             self.pending_transactions -= 1
         transaction.rollback(using=self.db_alias)
-        transaction.leave_transaction_management(using=self.db_alias)
+        if not DJANGO_17:
+            transaction.leave_transaction_management(using=self.db_alias)
 
     def rollback_transactions_dry_run(self):
         """
@@ -1044,7 +1051,7 @@ class DatabaseOperations(object):
             return
         while self.pending_transactions > 0:
             self.rollback_transaction()
-        if transaction.is_dirty(using=self.db_alias):
+        if not DJANGO_17 and transaction.is_dirty(using=self.db_alias):
             # Force an exception, if we're still in a dirty transaction.
             # This means we are missing a COMMIT/ROLLBACK.
             transaction.leave_transaction_management(using=self.db_alias)
@@ -1090,37 +1097,24 @@ class DatabaseOperations(object):
 
         created_models = []
         for model_name in model_names:
-            model = models.get_model(app_label, model_name)
-            if model:
-                created_models.append(model)
+            try:
+                model = models.get_model(app_label, model_name)
+            # Django 1.7 throws LookupError
+            except LookupError:
+                pass
+            else:
+                if model:
+                    created_models.append(model)
 
         if created_models:
-
-            if hasattr(dispatcher, "send"):
-                # Older djangos
-                dispatcher.send(signal=models.signals.post_syncdb, sender=app,
-                                app=app, created_models=created_models,
-                                verbosity=verbosity, interactive=interactive)
-            else:
-                if self._is_multidb():
-                    # Django 1.2+
-                    models.signals.post_syncdb.send(
-                        sender=app,
-                        app=app,
-                        created_models=created_models,
-                        verbosity=verbosity,
-                        interactive=interactive,
-                        db=self.db_alias,
-                    )
-                else:
-                    # Django 1.1 - 1.0
-                    models.signals.post_syncdb.send(
-                        sender=app,
-                        app=app,
-                        created_models=created_models,
-                        verbosity=verbosity,
-                        interactive=interactive,
-                    )
+            models.signals.post_syncdb.send(
+                sender=app,
+                app=app,
+                created_models=created_models,
+                verbosity=verbosity,
+                interactive=interactive,
+                db=self.db_alias,
+            )
 
     def mock_model(self, model_name, db_table, db_tablespace='',
                    pk_field_name='id', pk_field_type=models.AutoField,
